requirements.txt
FreeSimpleGUI
transformers
torch
Pillow
requests
beautifulsoup4